requirements.txt
fastapi
python-multipart
jinja2
uvicorn
httpx
pyjwt[crypto]
aiohttp
asyncio
sqlalchemy


